+++ data/resources/CHANGELOG.md
@@ -1,3 +1,7 @@
+June 11, 2021
+- Fix issue where invalid escape sequences would crash when printing to console (#322).
+- Fix /doors not searching text properly
+
 May 28, 2021
 - Fixed issue with keyboard input not working on popped-out windows
 - Implemented new console editor with support for word wrap and item links.